+++ packages/cadl-rlc-test/test/anomalyDetector/cadl-output/src/anomalyDetector.ts
@@ -10,9 +10,11 @@ export interface AnomalyDetectorClientOptions extends ClientOptions {
 }
 
 /**
- * Initialize a new instance of the class AnomalyDetectorClient class.
- * @param endpoint type: string
- * @param credentials type: KeyCredential
+ * Initialize a new instance of `AnomalyDetectorClient`
+ * @param endpoint type: string, Supported Cognitive Services endpoints (protocol and hostname, for example:
+ * https://westus2.api.cognitive.microsoft.com).
+ * @param credentials type: KeyCredential, uniquely identify client credential
+ * @param options type: AnomalyDetectorClientOptions, the parameter for all optional parameters
  */
 export default function createClient(
   endpoint: string,